+++ resources/[dg]/dg-financials/typescript/src/server/modules/bank/controllers/location.ts
@@ -3,5 +3,5 @@ import { Events } from '@dgx/server';
 import { setPlyLoc } from '../helpers/location';
 
 Events.onNet('financials:location:set', (src, locId: string | null) => {
-  setPlyLoc(source, locId);
+  setPlyLoc(src, locId);
 });